# api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: FinDoctor Client API
  version: 0.1.0
  summary: Client-facing API for financial accounts, ledger transactions, assets, liabilities, goals, tags, and analytics.
  license:
    name: Proprietary
    url: https://findoctor.example/legal/api-license
  description: |
    FinDoctor exposes user-owned financial data as explicit CRUD resources and keeps reference/admin data separate.

    Transfers are modeled as business objects connected to two underlying ledger transactions:
    one debit transaction on the source account and one credit transaction on the destination account.
    Creating, updating, or deleting a transfer must happen transactionally with those ledger rows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.findoctor.example
    description: Production placeholder
security:
  - bearerAuth: []
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/DateTimeFromQuery"
        - $ref: "#/components/parameters/DateTimeToQuery"
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/TransactionType"
        - name: merchantId
          in: query
          schema:
            type: string
            format: uuid
        - name: recurringTransactionId
          in: query
          schema:
            type: string
            format: uuid
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
    post:
      tags: [Transactions]
      summary: Create transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreate"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /transactions/{transactionId}:
    get:
      tags: [Transactions]
      summary: Get transaction
      operationId: getTransaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Transactions]
      summary: Update transaction
      operationId: updateTransaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Transactions]
      summary: Delete transaction
      operationId: deleteTransaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "204":
          description: Transaction deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /transactions/{transactionId}/tags:
    get:
      tags: [Transactions, Tags]
      summary: List transaction tags
      operationId: listTransactionTags
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    put:
      tags: [Transactions, Tags]
      summary: Replace all tags on a transaction
      operationId: replaceTransactionTags
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tagIds]
              properties:
                tagIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  uniqueItems: true
      responses:
        "200":
          description: Updated transaction tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
  /transactions/{transactionId}/tags/{tagId}:
    post:
      tags: [Transactions, Tags]
      summary: Attach tag to transaction
      operationId: addTransactionTag
      parameters:
        - $ref: "#/components/parameters/TransactionId"
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag attached
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Transactions, Tags]
      summary: Detach tag from transaction
      operationId: removeTransactionTag
      parameters:
        - $ref: "#/components/parameters/TransactionId"
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag detached
        "404":
          $ref: "#/components/responses/NotFound"
  /transfers:
    get:
      tags: [Transfers]
      summary: List transfers
      operationId: listTransfers
      parameters:
        - $ref: "#/components/parameters/DateTimeFromQuery"
        - $ref: "#/components/parameters/DateTimeToQuery"
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
          description: Match either source or destination account.
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: Paginated transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferPage"
    post:
      tags: [Transfers]
      summary: Create transfer and underlying transactions
      description: |
        Creates one transfer row and two linked transaction rows in a single database transaction.
        The source transaction is an expense-like transfer debit; the destination transaction is an income-like transfer credit.
      operationId: createTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferCreate"
      responses:
        "201":
          description: Created transfer with ledger transaction links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "422":
          $ref: "#/components/responses/ValidationError"
  /transfers/{transferId}:
    get:
      tags: [Transfers]
      summary: Get transfer
      operationId: getTransfer
      parameters:
        - $ref: "#/components/parameters/TransferId"
      responses:
        "200":
          description: Transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Transfers]
      summary: Update transfer and underlying transactions
      description: Updates the transfer row and synchronized amount, currency, datetime, account, and description fields on linked ledger transactions.
      operationId: updateTransfer
      parameters:
        - $ref: "#/components/parameters/TransferId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferUpdate"
      responses:
        "200":
          description: Updated transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Transfers]
      summary: Delete transfer and underlying transactions
      description: Deletes or reverses the transfer and both linked ledger transactions transactionally.
      operationId: deleteTransfer
      parameters:
        - $ref: "#/components/parameters/TransferId"
      responses:
        "204":
          description: Transfer and linked transaction rows deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DateTimeFromQuery:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    DateTimeToQuery:
      name: to
      in: query
      schema:
        type: string
        format: date-time
...
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 250
        default: 50
...
    TagId:
      name: tagId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TransferId:
      name: transferId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state or uniqueness constraints
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
  schemas:
...
    CurrencyCode:
      type: string
      minLength: 3
      maxLength: 3
      pattern: "^[A-Z]{3}$"
      examples: [USD, EUR, RUB]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
...
    MoneyAmount:
      type: string
      pattern: "^-?\\d+(\\.\\d{1,8})?$"
      description: Decimal amount encoded as a string to avoid floating point loss.
      examples: ["1250.00"]
...
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
...
    Tag:
      type: object
      required: [id, userId, name]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
...
    TagList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
    Transaction:
      allOf:
        - $ref: "#/components/schemas/TransactionCreate"
        - type: object
          required: [id, userId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            transferId:
              type: [string, "null"]
              format: uuid
              description: Present when this ledger row belongs to a transfer.
            transferLeg:
              type: [string, "null"]
              enum: [debit, credit, null]
            createdAt:
              type: string
              format: date-time
            tags:
              type: array
              items:
                $ref: "#/components/schemas/Tag"
    TransactionCreate:
      type: object
      required: [accountId, type, amount, currency, transactionDatetime]
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        transactionDatetime:
          type: string
          format: date-time
        description:
          type: [string, "null"]
        merchantId:
          type: [string, "null"]
          format: uuid
        merchantName:
          type: [string, "null"]
        geoLocation:
          type: [string, "null"]
        recurringTransactionId:
          type: [string, "null"]
          format: uuid
        externalId:
          type: [string, "null"]
        tagIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
...
    TransactionPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        meta:
          $ref: "#/components/schemas/PageMeta"
    TransactionType:
      type: string
      enum: [income, expense, transfer]
    TransactionUpdate:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        transactionDatetime:
          type: string
          format: date-time
        description:
          type: [string, "null"]
        merchantId:
          type: [string, "null"]
          format: uuid
        merchantName:
          type: [string, "null"]
        geoLocation:
          type: [string, "null"]
        recurringTransactionId:
          type: [string, "null"]
          format: uuid
        externalId:
          type: [string, "null"]
    Transfer:
      allOf:
        - $ref: "#/components/schemas/TransferCreate"
        - type: object
          required: [id, userId, fromTransactionId, toTransactionId]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            fromTransactionId:
              type: string
              format: uuid
              description: Underlying source-account ledger transaction.
            toTransactionId:
              type: string
              format: uuid
              description: Underlying destination-account ledger transaction.
            fromTransaction:
              $ref: "#/components/schemas/Transaction"
            toTransaction:
              $ref: "#/components/schemas/Transaction"
    TransferCreate:
      type: object
      required: [fromAccountId, toAccountId, amount, currency, transactionDatetime]
      properties:
        fromAccountId:
          type: string
          format: uuid
        toAccountId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        transactionDatetime:
          type: string
          format: date-time
        description:
          type: [string, "null"]
        categoryId:
          type: [string, "null"]
          format: uuid
          description: Optional category applied to both underlying transfer transactions.
        externalId:
          type: [string, "null"]
          description: Optional idempotency key for imported transfer creation.
    TransferUpdate:
      type: object
      properties:
        fromAccountId:
          type: string
          format: uuid
        toAccountId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        transactionDatetime:
          type: string
          format: date-time
        description:
          type: [string, "null"]
        categoryId:
          type: [string, "null"]
          format: uuid
    TransferPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        meta:
          $ref: "#/components/schemas/PageMeta"
...